fix(sonar): blockers, behavior-preserving cleanups, and parallel-parse race
SonarQube-driven code changes (core 663 + recovery tests green):
- fix(workspace): make TextDocument move assignment noexcept (cpp:S5018 blocker)
- fix(parser): compute Repetition shape facts in init_impl instead of lazily,
  removing a data race on the shared parser during parallel document builds
  (cpp:S8379)
- chore: behavior-preserving cleanups — S912 short-circuit->if, S6004 init-
  statements, S1186 empty-method comments, S5415 no-op std::move removal,
  S3574/S5827/S7172/S6228/S1110/S3230/S6222/S1905/S1066

+++ src/pegium/core/parser/RecoveryCandidate.hpp
@@ -189,8 +189,7 @@ static_assert(kFirstEditPenaltyStep ==
 /// (uint32) product cannot overflow for any realistic budget.
 [[nodiscard]] constexpr std::uint32_t
 faithfulness(const RecoveryKey &key) noexcept {
-  return key.editCost +
-         static_cast<std::uint32_t>(kFirstEditPenaltyStep) * key.editCount;
+  return key.editCost + kFirstEditPenaltyStep * key.editCount;
 }
 
 /// `firstEditOffset` penalised by edit cost: the score that axis 3
